# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Student Details - OpenAPI 3.0",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "email": "apiteam@swagger.io"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "1.0.11"
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "variables": {}
    }
  ],
  "tags": [
    {
      "name": "student-api",
      "description": "Everything about students"
    }
  ],
  "paths": {
    "/student/{rollNo}": {
      "get": {
        "tags": [
          "student-api"
        ],
        "summary": "Find student by RollNo",
        "description": "Returns a student",
        "operationId": "GetStudent",
        "parameters": [
          {
            "name": "rollNo",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/student"
                }
              }
            }
          },
          "400": {
            "description": "Invalid RollNo supplied"
          },
          "404": {
            "description": "Student Not Found"
          }
        }
      }
    },
    "/student": {
      "post": {
        "tags": [
          "student-api"
        ],
        "summary": "Add a new student to the store",
        "description": "Add a new student to the store",
        "operationId": "InsertStudent",
        "requestBody": {
          "description": "Create a new student in the store",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/student"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/student"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/student"
                }
              }
            }
          },
          "405": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/subject/{id}": {
      "get": {
        "tags": [
          "student-api"
        ],
        "summary": "Find a subject by id",
        "description": "Returns a subject",
        "operationId": "GetSubject",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/subject"
                }
              }
            }
          },
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "Student not found"
          }
        }
      }
    },
    "/subject": {
      "post": {
        "tags": [
          "student-api"
        ],
        "summary": "Add a new subject to the store",
        "description": "Add a new subject to the store",
        "operationId": "InsertSubject",
        "requestBody": {
          "description": "Create a new subject in the store",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/subject"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/subject"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/subject"
                }
              }
            }
          },
          "405": {
            "description": "Invalid input"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "student": {
        "type": "object",
        "properties": {
          "RollNo": {
            "type": "integer",
            "format": "int64",
            "example": 1
          },
          "Name": {
            "type": "string",
            "format": "string",
            "example": "Ankit"
          },
          "Age": {
            "type": "number",
            "format": "float",
            "example": 22
          }
        },
        "xml": {
          "name": "order"
        }
      },
      "subject": {
        "type": "object",
        "properties": {
          "Id": {
            "type": "integer",
            "format": "int64",
            "example": 10
          },
          "Name": {
            "type": "string",
            "format": "string",
            "example": "Science"
          }
        },
        "xml": {
          "name": "order"
        }
      },
      "enrollment": {
        "type": "object",
        "properties": {
          "Id": {
            "type": "integer",
            "format": "int64",
            "example": 10
          },
          "Name": {
            "type": "string",
            "format": "string",
            "example": "Marine"
          }
        },
        "xml": {
          "name": "order"
        }
      }
    }
  }
}